chaostooling-otel/pyproject.toml
[project]
name = "chaostooling-otel"
dynamic = ["version"]
description = "OpenTelemetry observability for Chaos Toolkit"
authors = [
    {name = "Morgan Wigge", email = "morgan@wigge.nu"},
    {name = "Chaos Toolkit", email = "contact@chaostoolkit.org"},
    {name = "Sylvain Hellegouarch", email = "sh@defuze.org"}
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: Freely Distributable",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation",
    "Programming Language :: Python :: Implementation :: CPython"
]
dependencies = [
   "chaostoolkit-lib>=1.42.1",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-semantic-conventions>=0.45b0",
    "opentelemetry-exporter-otlp>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=0.45b0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "opentelemetry-instrumentation>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-requests>=0.45b0",
    "opentelemetry-instrumentation-urllib3>=0.45b0",
    "opentelemetry-propagator-b3>=1.24.0",
    "opentelemetry-propagator-jaeger>=1.24.0",
    "opentelemetry-exporter-gcp-trace>=1.6.0",
    "opentelemetry-resourcedetector-gcp>=1.6.0a0",
    "opentelemetry-propagator-gcp>=1.6.0",
    "opentelemetry-exporter-prometheus>=0.45b0",
    "opentelemetry-sdk-extension-aws>=2.0.1",
    "opentelemetry-propagator-aws-xray>=1.0.1",
    "python-json-logger>=2.0.7",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "Apache-2.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
version = { source = "file", path = "chaosotel/__init__.py" }


[tool.pdm.scripts]
lint = {composite = ["ruff check chaosotel/", "isort --check-only --profile black chaosotel/", "black --check --diff chaosotel/", "mypy chaosotel/", "bandit -r chaosotel"]}
format = {composite = ["isort --profile black chaosotel/", "black chaosotel/", "ruff format chaosotel/ --fix"]}
test = {cmd = "pytest"}

[tool.black]
color = true
line-length = 80
target-version = ['py310']

[tool.mypy]
namespace_packages = true

[tool.ruff]
line-length = 80
target-version = "py310"
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "build",
    "dist",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false

[tool.isort]
line_length = 80
known_first_party = 'chaosotel'
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = "tests"
addopts = "-v -rxs --cov chaosotel --cov-report term-missing:skip-covered --cov-report xml -p no:warnings --junit-xml=./junit.xml"

[dependency-groups]
dev = [
    "coverage>=7.4.4",
    "pytest>=8.1.1",
    "pytest-cov>=5.0.0",
    "pytest-sugar>=1.0.0",
    "requests>=2.31.0",
    "requests-mock>=1.12.1",
    "ruff>=0.14.7",
    "isort",
    "black>=25.11.0",
    "mypy>=1.19.0",
    "bandit>=1.7.0",
]
